third_party/WebKit/LayoutTests/external/wpt/html/dom/dynamic-markup-insertion/opening-the-input-stream/015-expected.txt

This is a testharness.js-based test.
FAIL global scope unchanged assert_equals: expected 1 but got 2
PASS window object changed
PASS this is the window object
Harness: the test ran to completion.
